rework for high sierra
