+++ Crypt/Mechanisms/CryptMechanism.swift
@@ -138,4 +138,166 @@ class CryptMechanism: NSObject {
     _ = self.mechanism.pointee.fPlugin.pointee.fCallbacks.pointee.SetResult(
       mechanism.pointee.fEngine, AuthorizationResult.allow)
   }
+
+  func needToRestart() -> Bool {
+    os_log("Checking to see if we need to restart now because we may not be on APFS", log: CryptMechanism.log, type: .default)
+    let task = Process();
+    task.launchPath = "/usr/bin/fdesetup"
+    task.arguments = ["status"]
+    let pipe = Pipe()
+    task.standardOutput = pipe
+    task.launch()
+    let data = pipe.fileHandleForReading.readDataToEndOfFile()
+    guard let output: String = String(data: data, encoding: String.Encoding.utf8)
+      else { return true }
+    if ((output.range(of: "restart")) != nil) {
+      os_log("Looks like we need to restart...", log: CryptMechanism.log, type: .default)
+      return true
+    } else {
+      os_log("No restart needed.", log: CryptMechanism.log, type: .default)
+      return false
+    }
+  }
+  
+  // check if on 10.13+
+  func onHighSierraOrNewer() -> Bool {
+    os_log("Checking to see if on 10.13+", log: CryptMechanism.log, type: .default)
+    return ProcessInfo().isOperatingSystemAtLeast(OperatingSystemVersion.init(majorVersion: 10, minorVersion: 13, patchVersion: 0))
+  }
+  
+  // check authrestart capability
+  func checkAuthRestart() -> Bool {
+    let outPipe = Pipe.init()
+    let authRestartCheck = Process.init()
+    authRestartCheck.launchPath = "/usr/bin/fdesetup"
+    authRestartCheck.arguments = ["supportsauthrestart"]
+    authRestartCheck.standardOutput = outPipe
+    authRestartCheck.launch()
+    let outputData = outPipe.fileHandleForReading.availableData
+    let outputString = String(data: outputData, encoding: String.Encoding.utf8) ?? ""
+    if (outputString.range(of: "true") != nil) {
+      os_log("Authrestart capability is 'true', will authrestart as appropriate", log: CryptMechanism.log, type: .default)
+      return true
+    }
+    else {
+      os_log("Authrestart capability is 'false', reverting to standard reboot", log: CryptMechanism.log, type: .default)
+      return false
+    }
+  }
+  
+  // fdesetup Errors
+  private enum FileVaultError: Error {
+    case fdeSetupFailed(retCode: Int32)
+    case outputPlistNull
+    case outputPlistMalformed
+  }
+  
+  // Check if some information on filevault whether it's encrypted and if decrypting.
+  func getFVEnabled() -> (encrypted: Bool, decrypting: Bool) {
+    os_log("Checking the current status of FileVault..", log: CryptMechanism.log, type: .default)
+    let task = Process();
+    task.launchPath = "/usr/bin/fdesetup"
+    task.arguments = ["status"]
+    let pipe = Pipe()
+    task.standardOutput = pipe
+    task.launch()
+    let data = pipe.fileHandleForReading.readDataToEndOfFile()
+    guard let output: String = String(data: data, encoding: String.Encoding.utf8)
+      else { return (false, false) }
+    if ((output.range(of: "FileVault is On.")) != nil) {
+      os_log("Filevault is On...", log: CryptMechanism.log, type: .default)
+      return (true, false)
+    } else if (output.range(of: "Decryption in progress:") != nil) {
+      os_log("FileVault Decryption in progress...", log: CryptMechanism.log, type: .error)
+      return (true, true)
+    } else {
+      os_log("FileVault is not enabled...", log: CryptMechanism.log, type: .error)
+      return (false, false)
+    }
+  }
+
+  func enableFileVault(_ theSettings : NSDictionary, filepath : String) throws -> Bool {
+    os_log("Attempting to enable FileVault", log: CryptMechanism.log, type: .default)
+    let inputPlist = try PropertyListSerialization.data(fromPropertyList: theSettings,
+                                                        format: PropertyListSerialization.PropertyListFormat.xml, options: 0)
+    
+    let inPipe = Pipe.init()
+    let outPipe = Pipe.init()
+    let errorPipe = Pipe.init()
+    
+    let task = Process.init()
+    task.launchPath = "/usr/bin/fdesetup"
+    task.arguments = ["enable", "-outputplist", "-inputplist"]
+    
+    // check if we should do an authrestart on enablement
+    if checkAuthRestart() {
+      os_log("adding -authrestart flag at index 1 of our task arguments...", log: CryptMechanism.log, type: .default)
+      task.arguments?.insert("-authrestart", at: 1)
+    }
+    
+    // if there's an IRK, need to add the -keychain argument to keep us from failing.
+    let instKeyPath = "/Library/Keychains/FileVaultMaster.keychain"
+    if checkFileExists(path: instKeyPath) {
+      os_log("Appending -keychain to the end of our task arguments...", log: CryptMechanism.log, type: .default)
+      task.arguments?.append("-keychain")
+    }
+    
+    os_log("Running /usr/bin/fdesetup %{public}@", log: CryptMechanism.log, type: .default, String(describing: task.arguments))
+    
+    task.standardInput = inPipe
+    task.standardOutput = outPipe
+    task.standardError = errorPipe
+    task.launch()
+    inPipe.fileHandleForWriting.write(inputPlist)
+    inPipe.fileHandleForWriting.closeFile()
+    task.waitUntilExit()
+    
+    os_log("Trying to get output data", log: CryptMechanism.log, type: .default)
+    let outputData = outPipe.fileHandleForReading.readDataToEndOfFile()
+    outPipe.fileHandleForReading.closeFile()
+    
+    let errorData = errorPipe.fileHandleForReading.readDataToEndOfFile()
+    let errorMessage = String(data: errorData, encoding: .utf8)
+    errorPipe.fileHandleForReading.closeFile()
+    
+    if task.terminationStatus != 0 {
+      let termstatus = String(describing: task.terminationStatus)
+      os_log("fdesetup terminated with a NON-Zero exit status: %{public}@", log: CryptMechanism.log, type: .error, termstatus)
+      os_log("fdesetup Standard Error: %{public}@", log: CryptMechanism.log, type: .error, String(describing: errorMessage))
+      throw FileVaultError.fdeSetupFailed(retCode: task.terminationStatus)
+    }
+    
+    if outputData.count == 0 {
+      os_log("Found nothing in output data", log: CryptMechanism.log, type: .error)
+      throw FileVaultError.outputPlistNull
+    }
+    
+    var format : PropertyListSerialization.PropertyListFormat = PropertyListSerialization.PropertyListFormat.xml
+    let outputPlist = try PropertyListSerialization.propertyList(from: outputData,
+                                                                 options: PropertyListSerialization.MutabilityOptions(), format: &format)
+    
+    if (format == PropertyListSerialization.PropertyListFormat.xml) {
+      if outputPlist is NSDictionary {
+        os_log("Attempting to write key to: %{public}@", log: CryptMechanism.log, type: .default, String(describing: filepath))
+        _ = (outputPlist as! NSDictionary).write(toFile: filepath, atomically: true)
+      }
+      os_log("Successfully wrote key to: %{public}@", log: CryptMechanism.log, type: .default, String(describing: filepath))
+      return true
+    } else {
+      os_log("rotateRecoveryKey() Error. Format does not equal 'PropertyListSerialization.PropertyListFormat.xml'", log: CryptMechanism.log, type: .error)
+      throw FileVaultError.outputPlistMalformed
+    }
+  }
+  
+  func checkFileExists(path: String) -> Bool {
+    os_log("Checking to see if %{public}@ exists...", log: Check.log, type: .default, String(describing: path))
+    let fm = FileManager.default
+    if fm.fileExists(atPath: path) {
+      os_log("%{public}@ exists...", log: Check.log, type: .default, String(describing: path))
+      return true
+    } else {
+      os_log("%{public}@ doen NOT exists...", log: Check.log, type: .default, String(describing: path))
+      return false
+    }
+  }
 }
