+++ Package/Makefile
@@ -5,6 +5,7 @@ GITVERSION=$(shell ./build_no.sh)
 BUNDLE_VERSION=$(shell /usr/libexec/PlistBuddy -c "Print :CFBundleShortVersionString" "../Crypt/Info.plist")
 PACKAGE_VERSION=${BUNDLE_VERSION}.${GITVERSION}
 REVERSE_DOMAIN=com.grahamgilbert
+PACKAGE_NAME=${TITLE}
 PAYLOAD=\
 			pack-plugin\
 			pack-script-postinstall\
@@ -30,3 +31,10 @@ pack-checkin: l_Library
 	@sudo ${CP} FoundationPlist.py ${WORK_D}/Library/Crypt/FoundationPlist.py
 	@sudo chown -R root:wheel ${WORK_D}/Library/Crypt
 	@sudo chmod 755 ${WORK_D}/Library/Crypt/checkin
+
+dist: pkg
+	@sudo rm -f Distribution
+	python generate_dist.py
+	@sudo productbuild --distribution Distribution Crypt-${BUNDLE_VERSION}.pkg
+	@sudo rm -f Crypt.pkg
+	@sudo rm -f Distribution
